Implemented a function to check whether or not anyone is subscribed to a given triplet
owl_zephyr_destination_has_subscriptions pings the passed
<class,instance,user> tuple, and, on receiving a server ack, calls one
of two call-backs, depending on whether or not there is someone
subscribed.

The storage of callbacks is implemented with a new list in the
owl_global variable.  send_zephyr now has a variant send_zephyr_get_uid
that returns the z_uid of the sent notice.  send_ping now has a variant
send_ping_get_uid that returns the z_uid of the sent ping.