Added support to the default style for viewing chat state notifications
The flavor is based on jabber chat state notifications, XEP-0085.

Meta-messages are queued, and then deleted on update. The use of this
variant vs. display below the sepbar will be controlled by the
appropriate protocol-specific variable.

The default style is tweaked to display these messages intelligently.
Perhaps we should do something sane with other styles...