editwin callback for canceling the editwin
The code for editwin callbacks (called when the editwin is created) has
been extended to allow callbacks that get called when the editwin is
canceled.  The editwin callbacks now take a second boolean parameter
that tells the callback whether or not the editwin was canceled.

A new function, start_edit, was created as the preferred interface.

The perl interface to the editwin is backwards compatible; the start_*
commands take an optional third parameter, and the callback is only
called on cancel if this parameter is true.